Correcion: Por si se reparte +1300 desayunos

+++ database/migrations/2025_11_12_125638_create_detalle_registro_diarios_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('detalle_registro_diarios', function (Blueprint $table) {
             $table->id();
             $table->foreignId('receta_id')->constrained('recetas')->onDelete('cascade');
-            $table->integer('cantidad_servido');
+            $table->bigInteger('cantidad_servido');
             $table->timestamps();
         });
     }
